--- org_units/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class OrgUnitsConfig(AppConfig):
-    name = 'org_units'
