+++ programs/models.py
@@ -11,7 +11,7 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from rest_framework.authtoken.models import Token
 
-from org_units.models import Unit
+from units.models import Unit
 
 # Create your models here.
