+++ settings.py
@@ -43,7 +43,7 @@ INSTALLED_APPS = [
     'teledata',
     'research',
     'images',
-    'org_units',
+    'units',
     'core'
 ]
