+++ teledata/admin.py
@@ -41,7 +41,7 @@ class OrganizationAdmin(admin.ModelAdmin):
         return mark_safe(
             '<a href="{0}">{1}</a>'.format(
                 reverse(
-                    'admin:org_units_unit_change',
+                    'admin:units_unit_change',
                     args=(obj.unit.pk,)
                 ),
                 obj.unit.name
@@ -73,7 +73,7 @@ class DepartmentAdmin(admin.ModelAdmin):
         return mark_safe(
             '<a href="{0}">{1}</a>'.format(
                 reverse(
-                    'admin:org_units_unit_change',
+                    'admin:units_unit_change',
                     args=(obj.unit.pk,)
                 ),
                 obj.unit.name
